Recruitment and Training
Create an Automation System of Recruitment and Training process for the
Organization to reduce man work and efforts, enhance productivity,
reduce wastage and quality Control. The main aim is to develop in a such
a way that manages Training Modules of the industry as to a provide
proper navigation and better security.